+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Fix broken codesign option for bundle dependency [#1104](https://github.com/yonaskolb/XcodeGen/pull/1104) @kateinoigakukun
 - Ensure fileTypes are mapped to JSON value [#1112](https://github.com/yonaskolb/XcodeGen/pull/1112) @namolnad
 - Fix platform filter for package dependecies [#1123](https://github.com/yonaskolb/XcodeGen/pull/1123) @raptorxcz
-- Fix Xcode 13 build [#1127](https://github.com/yonaskolb/XcodeGen/issues/1127) @raptorxcz
+- Fix Xcode 13 build [#1130](https://github.com/yonaskolb/XcodeGen/issues/1127) @raptorxcz @mthole
 
 [Commits](https://github.com/yonaskolb/XcodeGen/compare/2.24.0...2.25.0)